requirements.txt
caldav
pillow
vobject
spidev
gpiozero
